requirements.txt
streamlit
google-generativeai
streamlit-mic-recorder
gTTS